% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecgca102}
\alias{ecgca102}
\title{Time Series Data of Electrocardiogram}
\format{a numeric vector}
\source{
Non-Invasive Fetal Electrocardiogram Database (\url{https://doi.org/10.13026/C2X30H})
}
\usage{
data("ecgca102")
}
\description{
This data is a part of ecgca102.edf file of Non-Invasive Fetal Electrocardiogram Database.
}
\references{
Goldberger AL, Amaral LAN, Glass L, Hausdorff JM, Ivanov PCh, Mark RG, Mietus JE, Moody GB, Peng C-K, Stanley HE. PhysioBank, PhysioToolkit, and PhysioNet: Components of a New Research Resource for Complex Physiologic Signals. Circulation 101(23):e215-e220 [Circulation Electronic Pages; http://circ.ahajournals.org/cgi/content/full/101/23/e215]; 2000 (June 13).
}
\keyword{datasets}
