% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{package}
\name{scorepeak}
\alias{scorepeak}
\alias{scorepeak-package}
\title{scorepeak: Peak Functions for Peak Detection in Univariate Time Series}
\description{
scorepeak provides peak functions and its building blocks.
Peak functions enable us to detect peaks.
}
